rag_chatbot/requirements.txt
streamlit==1.42.0
langchain==0.3.19
langchain-community==0.3.18
langchain-text-splitters==0.3.6
chromadb==0.5.20
pypdf==5.1.0
sentence-transformers==3.3.1
torch>=2.1.0
transformers>=4.40.0
python-dotenv==1.0.1